Домашнее задание к занятию «3.2. Generics и коллекции», доработка по замечаниям: - сделал дженерик интерфейс, - имплементацию NoteService.kt и тесты к ней, - надо сделать ещё имплементацию CommentService.kt и тесты к ней:    - сделал add и тесты к нему

+++ src/main/kotlin/ru/netology/hwKotlin/Comment.kt
@@ -3,7 +3,7 @@ package ru.netology.hwKotlin
 data class Comment(
     val id: Long,
     val fromID: Int,
-    val targetId: Int,
+    val targetId: Long,
     val date: Int,
     val text: String,
     val donut: Donut,
